Fixing broken compatibility (#828)
* Generate raw chainspecs in script

* Make compatibility work again

+++ finality-aleph/src/abft/current.rs
@@ -17,7 +17,7 @@ use crate::{
 };
 
 /// Version of the current abft
-pub const VERSION: u32 = 2;
+pub const VERSION: u16 = 2;
 
 pub fn run_member<
     B: Block,
